Added File to Help Get Started
This file generates a folder that contains each unique video, roughly 7.4k, and stores the frames from each video together. Methods exist to grab these randomly, so there are no issues. However, this helps with data optimization, as well as speed of compilation. Should only need to be run once to generate the folder. 

The code after generating the folder serves to demonstrate some basic ways you can pull data from the files, such as the x, y and z coordinate arrays, the specific frame number of the frame from a video, and the true emotion label.